Fixed issue with assuming each colour value is a 4 byte int rather than the byte it is supposed to be.

+++ src/bitmaps/writer/writer.cpp
@@ -76,7 +76,7 @@ ByteArray Bitmaps::ImageWriter::convertToBytes(const Bitmaps::Colours& colours)
   ByteArray byteArray {};
 
   for (auto colourAndValue : colours)
-    byteArray.add(colourAndValue.second);
+    byteArray.add((uint8_t) colourAndValue.second);
 
   return byteArray;
 }
